Refactor entity package moved to model package
