+++ src/main/java/com/study/loge/betchat/model/entity/Participate.java
@@ -1,4 +1,4 @@
-package com.study.loge.betchat.entity;
+package com.study.loge.betchat.model.entity;
 
 import lombok.*;
