+++ src/main/java/com/study/loge/betchat/model/entity/Room.java
@@ -1,4 +1,4 @@
-package com.study.loge.betchat.entity;
+package com.study.loge.betchat.model.entity;
 
 import lombok.*;
 
@@ -6,30 +6,28 @@ import javax.persistence.*;
 import java.time.LocalDateTime;
 import java.util.List;
 
-// User Entity를 표현하기 위한 Entity 클래스입니다.
+// Room Entity를 표현하기 위한 Entity 클래스입니다.
 @Entity
 @AllArgsConstructor
 @NoArgsConstructor
 @Getter
 @Setter
 @Builder
-public class User {
+public class Room {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
-    private String userName;
-    private String userKey;
+    private String roomKey;
     private LocalDateTime createdAt;
 
-    @OneToMany(fetch = FetchType.LAZY, mappedBy = "user")
+    @OneToMany(fetch = FetchType.LAZY, mappedBy = "room")
     private List<Participate> participate;
 
     @Override
     public String toString() {
-        return "User{" +
+        return "Room{" +
                 "id=" + id +
-                ", userName='" + userName + '\'' +
-                ", userKey='" + userKey + '\'' +
+                ", roomKey='" + roomKey + '\'' +
                 ", createdAt=" + createdAt +
                 '}';
     }
