+++ src/main/java/com/study/loge/betchat/model/entity/User.java
@@ -1,4 +1,4 @@
-package com.study.loge.betchat.entity;
+package com.study.loge.betchat.model.entity;
 
 import lombok.*;
 
@@ -6,28 +6,30 @@ import javax.persistence.*;
 import java.time.LocalDateTime;
 import java.util.List;
 
-// Room Entity를 표현하기 위한 Entity 클래스입니다.
+// User Entity를 표현하기 위한 Entity 클래스입니다.
 @Entity
 @AllArgsConstructor
 @NoArgsConstructor
 @Getter
 @Setter
 @Builder
-public class Room {
+public class User {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
-    private String roomKey;
+    private String userName;
+    private String userKey;
     private LocalDateTime createdAt;
 
-    @OneToMany(fetch = FetchType.LAZY, mappedBy = "room")
+    @OneToMany(fetch = FetchType.LAZY, mappedBy = "user")
     private List<Participate> participate;
 
     @Override
     public String toString() {
-        return "Room{" +
+        return "User{" +
                 "id=" + id +
-                ", roomKey='" + roomKey + '\'' +
+                ", userName='" + userName + '\'' +
+                ", userKey='" + userKey + '\'' +
                 ", createdAt=" + createdAt +
                 '}';
     }
