+++ src/main/java/com/study/loge/betchat/repository/ParticipateRepository.java
@@ -1,6 +1,6 @@
 package com.study.loge.betchat.repository;
 
-import com.study.loge.betchat.entity.Participate;
+import com.study.loge.betchat.model.entity.Participate;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
