+++ src/main/java/com/study/loge/betchat/repository/RoomRepository.java
@@ -1,6 +1,6 @@
 package com.study.loge.betchat.repository;
 
-import com.study.loge.betchat.entity.Room;
+import com.study.loge.betchat.model.entity.Room;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
