+++ src/main/java/com/study/loge/betchat/utils/validation/messageheader/RoomJoinValidationChecker.java
@@ -1,6 +1,6 @@
 package com.study.loge.betchat.utils.validation.messageheader;
 
-import com.study.loge.betchat.entity.Room;
+import com.study.loge.betchat.model.entity.Room;
 import com.study.loge.betchat.utils.exception.RoomJoinException;
 import com.study.loge.betchat.model.MessageHeader;
 import com.study.loge.betchat.model.request.RoomJoinRequest;
